Rename webpack for karma and storybook to match existing offline
